Updated code to reverse string
The initial code were having a two major issues, that is in one line a semicolon was missing and the other was with the for loop in which the string was actually being reversed. I tried to correct the loop conditions by updating it so that even the odd length strings have no issues in reversal. Since the time complexity is of order of O(n) I didn't think it required optimization.

Note: I have registered with a different email id in the form i.e. , ap23csb0f47@student.nitw.ac.in and am using this email id for Git Hub.